requirements.txt
python-telegram-bot==21.2
httpx==0.27.0


